requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.3
msgpack==1.0.5
psutil==5.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
pyzmq==25.1.0
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.6.3
urllib3==1.26.14
Werkzeug==2.3.6
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
